rename to ec-cli
