+++ .copier-answers.yml
@@ -5,10 +5,10 @@ author_email: giles.knap@gmail.com
 author_name: Giles Knap
 component_owner: group:default/sscc
 description: CLI for deploying and managing epics containers IOCs and services
-distribution_name: epics-containers-cli
+distribution_name: ec-cli
 docker: false
 docs_type: README
 git_platform: github.com
 github_org: epics-containers
-package_name: epics_containers_cli
-repo_name: epics-containers-cli
+package_name: ec_cli
+repo_name: ec-cli
