+++ README.rst
@@ -1,4 +1,4 @@
-epics_containers_cli
+ec_cli
 =============================================================================
 
 |code_ci| |docs_ci| |coverage| |pypi_version| |license|
@@ -7,10 +7,10 @@ This is where you should write a short paragraph that describes what your module
 how it does it, and why people should use it.
 
 ============== ==============================================================
-PyPI           ``pip install epics-containers-cli``
-Source code    https://github.com/epics-containers/epics-containers-cli
+PyPI           ``pip install ec-cli``
+Source code    https://github.com/epics-containers/ec-cli
 
-Releases       https://github.com/epics-containers/epics-containers-cli/releases
+Releases       https://github.com/epics-containers/ec-cli/releases
 ============== ==============================================================
 
 This is where you should put some images or code snippets that illustrate
@@ -19,28 +19,28 @@ introductory code here:
 
 .. code-block:: python
 
-    from epics_containers_cli import __version__
+    from ec_cli import __version__
 
-    print(f"Hello epics_containers_cli {__version__}")
+    print(f"Hello ec_cli {__version__}")
 
 Or if it is a commandline tool then you might put some example commands here::
 
-    $ python -m epics_containers_cli --version
+    $ python -m ec_cli --version
 
-.. |code_ci| image:: https://github.com/epics-containers/epics-containers-cli/actions/workflows/code.yml/badge.svg?branch=main
-    :target: https://github.com/epics-containers/epics-containers-cli/actions/workflows/code.yml
+.. |code_ci| image:: https://github.com/epics-containers/ec-cli/actions/workflows/code.yml/badge.svg?branch=main
+    :target: https://github.com/epics-containers/ec-cli/actions/workflows/code.yml
     :alt: Code CI
 
-.. |docs_ci| image:: https://github.com/epics-containers/epics-containers-cli/actions/workflows/docs.yml/badge.svg?branch=main
-    :target: https://github.com/epics-containers/epics-containers-cli/actions/workflows/docs.yml
+.. |docs_ci| image:: https://github.com/epics-containers/ec-cli/actions/workflows/docs.yml/badge.svg?branch=main
+    :target: https://github.com/epics-containers/ec-cli/actions/workflows/docs.yml
     :alt: Docs CI
 
-.. |coverage| image:: https://codecov.io/gh/epics-containers/epics-containers-cli/branch/main/graph/badge.svg
-    :target: https://codecov.io/gh/epics-containers/epics-containers-cli
+.. |coverage| image:: https://codecov.io/gh/epics-containers/ec-cli/branch/main/graph/badge.svg
+    :target: https://codecov.io/gh/epics-containers/ec-cli
     :alt: Test Coverage
 
-.. |pypi_version| image:: https://img.shields.io/pypi/v/epics-containers-cli.svg
-    :target: https://pypi.org/project/epics-containers-cli
+.. |pypi_version| image:: https://img.shields.io/pypi/v/ec-cli.svg
+    :target: https://pypi.org/project/ec-cli
     :alt: Latest PyPI version
 
 .. |license| image:: https://img.shields.io/badge/License-Apache%202.0-blue.svg
