+++ docs/index.rst
@@ -14,13 +14,13 @@ The documentation is split into 2 sections:
         :link: user/index
         :link-type: doc
 
-        The User Guide contains documentation on how to install and use epics-containers-cli.
+        The User Guide contains documentation on how to install and use ec-cli.
 
     .. grid-item-card:: :material-regular:`code;4em`
         :link: developer/index
         :link-type: doc
 
-        The Developer Guide contains documentation on how to develop and contribute changes back to epics-containers-cli.
+        The Developer Guide contains documentation on how to develop and contribute changes back to ec-cli.
 
 .. toctree::
     :hidden:
