+++ docs/user/reference/api.rst
@@ -1,14 +1,14 @@
 API
 ===
 
-.. automodule:: epics_containers_cli
+.. automodule:: ec_cli
 
-    ``epics_containers_cli``
+    ``ec_cli``
     -----------------------------------
 
-This is the internal API reference for epics_containers_cli
+This is the internal API reference for ec_cli
 
-.. data:: epics_containers_cli.__version__
+.. data:: ec_cli.__version__
     :type: str
 
     Version number as calculated by https://github.com/pypa/setuptools_scm
