+++ src/ec_cli/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import version  # noqa
 
-__version__ = version("epics-containers-cli")
+__version__ = version("ec-cli")
 del version
 
 __all__ = ["__version__"]
