+++ src/ec_cli/__main__.py
@@ -2,8 +2,8 @@ from typing import Optional
 
 import typer
 
-import epics_containers_cli.globals as globals
-from epics_containers_cli.ioc.ioc_cli import cli
+import ec_cli.globals as globals
+from ec_cli.ioc.ioc_cli import cli
 
 from . import __version__
 from .logging import init_logging
@@ -64,6 +64,6 @@ def main(
 
 
 # test with:
-#     python -m epics_containers_cli
+#     python -m ec_cli
 if __name__ == "__main__":
     cli()
