+++ src/ec_cli/globals.py
