+++ src/ec_cli/ioc/ioc_cli.py
@@ -4,19 +4,19 @@ from pathlib import Path
 import typer
 from natsort import natsorted
 
-import epics_containers_cli.globals as globals
-from epics_containers_cli.autocomplete import (
+import ec_cli.globals as globals
+from ec_cli.autocomplete import (
     all_iocs,
     avail_IOCs,
     avail_versions,
     force_plain_completion,
     running_iocs,
 )
-from epics_containers_cli.git import create_ioc_graph
-from epics_containers_cli.ioc.k8s_commands import IocK8sCommands
-from epics_containers_cli.ioc.local_commands import IocLocalCommands
-from epics_containers_cli.logging import log
-from epics_containers_cli.utils import cleanup_temp, drop_ioc_path
+from ec_cli.git import create_ioc_graph
+from ec_cli.ioc.k8s_commands import IocK8sCommands
+from ec_cli.ioc.local_commands import IocLocalCommands
+from ec_cli.logging import log
+from ec_cli.utils import cleanup_temp, drop_ioc_path
 
 cli = typer.Typer(pretty_exceptions_show_locals=False)
