+++ tests/test_cli.py
@@ -1,9 +1,9 @@
 import subprocess
 import sys
 
-from epics_containers_cli import __version__
+from ec_cli import __version__
 
 
 def test_cli_version():
-    cmd = [sys.executable, "-m", "epics_containers_cli", "--version"]
+    cmd = [sys.executable, "-m", "ec_cli", "--version"]
     assert subprocess.check_output(cmd).decode().strip() == __version__
